+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 # UNRELEASED
 
 * AADServicePrincipal
-  * Streamlined logic to translate AppId from a display name in the
-    Set-TargetResource function.
+  * Streamlined logic to translate AppId from a display name.
 * IntuneDeviceConfigurationPolicyAndroidDeviceOwner
   * Fixed an issue where empty complex objects result in a broken export.
     FIXES [#5863](https://github.com/microsoft/Microsoft365DSC/issues/5863)